Split ImageHelpers into different classes
